+++ assets/components/productPage/productPagePeriodForm/productPagePeriodForm.scss
@@ -1,4 +1,4 @@
-.weekly-form {
+.component-product-page-period-form {
   margin-bottom: $gu-v-spacing * 2;
   @include mq($from: tablet) {
     display: flex;
@@ -6,7 +6,7 @@
   }
 }
 
-.weekly-form > .weekly-form__item {
+.component-product-page-period-form > .component-product-page-period-form__item {
   flex: 0 0 25%;
   padding: 0 $gu-h-spacing/2;
   box-sizing: border-box;
@@ -16,7 +16,7 @@
   }
 }
 
-.weekly-form__info {
+.component-product-page-period-form__info {
   padding-top: $gu-v-spacing * 2;
   padding-bottom: $gu-v-spacing * 2.5;
   color: gu-colour(garnett-neutral-5);
@@ -32,7 +32,7 @@
   }
 }
 
-.weekly-form__cta {
+.component-product-page-period-form__cta {
   @include mq($until: tablet) {
     transition: .4s;
     transform: translateY(0);
