+++ assets/pages/weekly-subscription-landing/components/weeklyForm.jsx
@@ -5,19 +5,39 @@
 import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
 
-import { type WeeklyBillingPeriod } from 'helpers/subscriptions';
+import { type CountryGroupId } from 'helpers/internationalisation/countryGroup';
+import { currencies, detect } from 'helpers/internationalisation/currency';
+import { type WeeklyBillingPeriod, getWeeklyProductPrice } from 'helpers/subscriptions';
+import { type Action } from 'components/productPage/productPagePeriodForm/productPagePeriodFormActions';
+import ProductPagePeriodForm from 'components/productPage/productPagePeriodForm/productPagePeriodForm';
 
-import { billingPeriods, type State } from '../weeklySubscriptionLandingReducer';
+import { type State } from '../weeklySubscriptionLandingReducer';
 import { redirectToWeeklyPage, setPeriod } from '../weeklySubscriptionLandingActions';
-import { type Action } from './productPagePeriodFormActions';
-
-import ProductPagePeriodForm from './productPagePeriodForm';
 
 
-// ---- Types ----- //
+// ---- Periods ----- //
 
+const getPrice = (countryGroupId: CountryGroupId, period: WeeklyBillingPeriod) => [
+  currencies[detect(countryGroupId)].extendedGlyph,
+  getWeeklyProductPrice(countryGroupId, period),
+].join('');
 
-// ----- Render ----- //
+export const billingPeriods = {
+  sixweek: {
+    title: '6 for 6',
+    offer: 'Introductory offer',
+    copy: (countryGroupId: CountryGroupId) => `${getPrice(countryGroupId, 'sixweek')} for the first 6 issues (then ${getPrice(countryGroupId, 'quarter')} quarterly)`,
+  },
+  quarter: {
+    title: 'Quarterly',
+    copy: (countryGroupId: CountryGroupId) => `${getPrice(countryGroupId, 'quarter')} every 3 months`,
+  },
+  year: {
+    title: 'Annually',
+    copy: (countryGroupId: CountryGroupId) => `${getPrice(countryGroupId, 'year')} every 12 months`,
+  },
+};
+
 
 // ----- State/Props Maps ----- //
 
@@ -28,7 +48,6 @@ const mapStateToProps = (state: State) => {
       title: billingPeriods[k].title,
       copy: billingPeriods[k].copy(state.common.internationalisation.countryGroupId),
       offer: billingPeriods[k].offer || null,
-
     };
   });
 
@@ -44,6 +63,7 @@ const mapDispatchToProps = (dispatch: Dispatch<Action<WeeklyBillingPeriod>>) => 
   setPeriodAction: bindActionCreators(setPeriod, dispatch),
   onSubmit: bindActionCreators(redirectToWeeklyPage, dispatch),
 });
+
 
 // ----- Exports ----- //
